Lib

Documentation

gadtExpr :: ()

GADT implementation of logical expressions

data Expr a where
  Lit :: Bool -> Expr Bool
  And :: Expr Bool -> Expr Bool -> Expr Bool
  Or  :: Expr Bool -> Expr Bool -> Expr Bool
  Xor :: Expr Bool -> Expr Bool -> Expr Bool
  Not :: Expr Bool -> Expr Bool

data And a

Constructors

And a a

data Or a

Constructors

Or a a

data Xor a

Constructors

Xor a a

data Not a

Constructors

Not a

type AndF = Free And

simplAndEqArgs :: Eq a => And a -> Either a (And a)

Example lowest-level optimization:

 x `And` x = x

liftCollapse :: (Functor f, c (Free f a)) => proxy c -> (forall t. c t => f t -> Either t (f t)) -> Free f a -> FreeT f (Either a) a

Lift with the given function, collapsing and reapplying when Left is returned

simplAndEqArgsF :: Eq a => AndF a -> AndF a

reduceSimpl :: Functor f => FreeT f (Either a) a -> Free f a

type OrF = Free Or

type XorF = Free Xor

type NotF = Free Not

type ExprF = Free (And :+: (Or :+: (Xor :+: Not)))

getAnd :: ExprF a -> Either (AndF a) (ExprF (AndF a))

getAndF :: ExprF a -> ExprF (AndF a)

fromAndF :: AndF a -> ExprF a

partAnd :: ExprF (AndF a) -> ExprF (ExprF (AndF a))

type AndO = Free (Or :+: (Xor :+: Not))

partDartPver :: ()

Part and dart notes

 part :: [] . Maybe -> [] . (Maybe . NonEmpty) . Maybe
 part ~ map Just . groupOn isJust

 dart :: [] . (Maybe . NonEmpty) . Maybe -> [] . Maybe . NonEmpty
 dart ~ map (fmap (fmap fromJust))

 pver :: [] . Maybe -> [] . Maybe . []
 pver ~ fmap (fmap toList) . dart . part

An expression type is (up to) the recursive completion of the exclusive or (non-inclusive sum) of two expression types.

 A functor is (up to) isomorphic to the recursive completion of the xor of two other functors.
   f ~ g xor h
 With the boundary language introduced before, we may bound the "leafy" subtrees that completely consist of one functor.
   This gives us: f -> f . g (|| f -> f . h)
 Next, we split out the trees that contain those, and only the other functor:
   f . g -> f . h . g
 Finally we repeat the process for the original "sub" functor:
   f . h . g -> f . g . h . g
 Alternatively, we can think of the pattern: "^.*GHG$", where G, H represent the matcher of a single element of a path in g, h, resp.

pushWithSplits :: ()

 Expr = ExprL xor ExprR
 splitL :: Expr -> Expr . ExprL
 splitR :: Expr -> Expr . ExprR
 splitRL :: Expr . ExprR -> Expr . (ExprR . ExprL) . ExprR

 splitLR :: Expr . ExprL -> Expr . (ExprL . ExprR) . ExprL
 assoc :: Expr . (ExprL . ExprR) . ExprL -> Expr . ExprL . (ExprR . ExprL)
 joinRL :: ExprR . ExprL -> Expr
 push :: Expr . ExprL -> Expr . ExprL . Expr
 push = fmap (fmap joinRL) . assoc . split

 pushLR :: Expr . ExprL -> Expr . ExprL . ExprR . ExprL
 pushLR = assoc . split

 part :: ExprF . AndF -> ExprF . (AndF . AndO) . AndF

 dart :: (AndF . AndO) . AndF -> AndF . (AndO . AndF)

 push :: ExprF

 A "leafy" Subtree of And's is: AndF a -> AndF (Pure a)
   A subexpression of Just And's

 A "branchy" Subtree of And's is AndF (not Pure) -> AndF (ExprF a)
   A subexpression of And's ending is non-And, then And, etc..

logicalForms :: ()

Several logical forms:

 Or (And a)
 (x && y) || z => (x || y) && (y || z)
 (x y z w->(x && y) || (z && w) == (x || z) && (x || w) && (y || z) && (y || w))

 DNF
 (x && y) || (z && w)

 CNF
 (w ∨ x) ∧ (w ∨ y) ∧ (x ∨ z) ∧ (y ∨ z)

 (x && y) xor (z && w)

 CNF
 (¬w ∨ ¬x ∨ ¬y ∨ ¬z) ∧ (w ∨ x) ∧ (w ∨ y) ∧ (x ∨ z) ∧ (y ∨ z)

partAndo :: ExprF (AndF a) -> ExprF (AndO (AndF a))

undefined

dartAndo :: ExprF (AndO (AndF a)) -> ExprF (AndF (AndO a))

dartAnd :: ExprF (ExprF (AndF a)) -> ExprF (AndF (ExprF a))

undefined

 dartAnd (Pure x) = _ x
 dartAnd (Free (L1 (And x y))) = _ x
 dartAnd (Free (R1 (L1 x))) = _ x
 dartAnd (Free (R1 (R1 x))) = _ x
 dartAnd (Free (R1 (R1 (L1 (Xor x y))))) = _ x y
 dartAnd (Free (R1 (R1 (R1 (Not x))))) = _ x

 dartAnd (Pure (Pure x)) = Pure $ Pure $ x
 dartAnd (Pure (Free (L1 (And x y)))) =

pushAnd :: ExprF (AndF a) -> ExprF (AndF (ExprF a))

qartAnd :: ExprF (ExprF (AndF a)) -> ExprF (AndF a)

bartAnd :: ExprF (AndF (ExprF a)) -> ExprF (ExprF (AndF a))

pullAnd :: ExprF (AndF (ExprF a)) -> ExprF (AndF a)

simplAndEqLoop :: Eq a => ExprF a -> ExprF a

Simplify recursively

 simplAndEqLoop xs =
   (>>= fromAndF) . pullAnd . fmap simplAndEqArgsF . pushAnd . getAndF $ xs

Orphan instances

(Eq1 f, Eq1 g) => Eq1 (f :+: g)

Instance details

Methods

liftEq :: (a -> b -> Bool) -> (f :+: g) a -> (f :+: g) b -> Bool